refactor: move component to its own directory
