+++ src/index.tsx
@@ -3,7 +3,7 @@ import * as ReactDOM from 'react-dom'
 import './index.scss'
 
 import App from './App'
-import ErrorBoundaries from './shared/components/ErrorBoundary'
+import ErrorBoundaries from './shared/components/ErrorBoundaries/ErrorBoundaries'
 
 const app = (
 	<ErrorBoundaries>
